feat: support of auth headers from gateway

+++ app/routers/v1/fhir/localization_list.py
@@ -9,6 +9,7 @@ from fastapi.params import Query
 from app.dependencies import (
     get_localization_list_service,
 )
+from app.exceptions.fhir_exception import UnauthorizedAction
 from app.models.fhir.resources.data import (
     DATA_DOMAIN_SYSTEM,
     DEVICE_SYSTEM,
@@ -25,6 +26,7 @@ from app.models.fhir.resources.operation_outcome.resource import OperationOutcom
 from app.models.response import DeleteResponse, FHIRJSONResponse
 from app.models.ura import UraNumber
 from app.routers.v1.fhir import SUBJECT_IDENTIFIER_PARAM
+from app.services.auth.auth_context import AuthContextService, RequestedAction
 from app.services.fhir.localization_list import LocalizationListService
 
 logger = logging.getLogger(__name__)
@@ -153,6 +155,10 @@ def create(
     request: Request,
     service: Annotated[LocalizationListService, Depends(get_localization_list_service)],
 ) -> Any:
+    valid_action = AuthContextService.validate_action(request.state.auth, RequestedAction.MODIFYING)
+    if not valid_action:
+        raise UnauthorizedAction(request.state.auth.role, RequestedAction.MODIFYING.value)
+
     authorized_ura: UraNumber = request.state.auth.ura_number
     return service.create(data, authorized_ura)
 
@@ -317,6 +323,10 @@ def query(
     params: Annotated[LocalizationListParams, Query()],
     service: Annotated[LocalizationListService, Depends(get_localization_list_service)],
 ) -> Any:
+    valid_action = AuthContextService.validate_action(request.state.auth, RequestedAction.LOCALIZING)
+    if not valid_action:
+        raise UnauthorizedAction(request.state.auth.role, RequestedAction.LOCALIZING.value)
+
     authorized_ura = request.state.auth.ura_number
     return service.query(params, authorized_ura)
 
@@ -342,6 +352,10 @@ def delete(
     request: Request,
     service: Annotated[LocalizationListService, Depends(get_localization_list_service)],
 ) -> Any:
+    valid_action = AuthContextService.validate_action(request.state.auth, RequestedAction.MODIFYING)
+    if not valid_action:
+        raise UnauthorizedAction(request.state.auth.role, RequestedAction.MODIFYING.value)
+
     authorized_ura: UraNumber = request.state.auth.ura_number
     outcome, status_code = service.delete(id, authorized_ura)
     return FHIRJSONResponse(
@@ -371,6 +385,10 @@ def delete_for_query(
     params: Annotated[LocalizationListParams, Query()],
     service: Annotated[LocalizationListService, Depends(get_localization_list_service)],
 ) -> Any:
+    valid_action = AuthContextService.validate_action(request.state.auth, RequestedAction.MODIFYING)
+    if not valid_action:
+        raise UnauthorizedAction(request.state.auth.role, RequestedAction.MODIFYING.value)
+
     authenticated_ura = request.state.auth.ura_number
     outcome, status_code = service.delete_by_query(params, authenticated_ura)
